refactor(admin): extract installFakeIntersectionObserver across editor tests
Five test files had grown nearly-identical copies of the IntersectionObserver test harness
(LazyMount / PatternsSection / BlockScopePicker / InsertableEntryRow / SlashMenuPanel), each
~40 lines of FakeObserver class + intersect() helper. PatternsSection's variant captured the
observed target; the other four synthesised one. Senpai flagged the duplication twice as the
cost-benefit kept getting worse with each new consumer.

`installFakeIntersectionObserver()` is the file-scope fixture; each test file calls it once
at module top-level and destructures the returned `intersect`. The harness captures the
real observed target (PatternsSection's behaviour) so future per-target assertions still
work; the four sites that didn't assert on target keep working without changes.

Senpai pre-commit findings folded in:
- `observations` is `const` + reset via `.length = 0`. Eliminates the autofix vector where
  `prefer-const` could silently break per-test isolation by converting a load-bearing `let`.
- Comment on `installFakeIntersectionObserver` notes the harness is unsafe under
  `test.concurrent` — `observations` is file-scoped, so parallel tests would race.

The deeper interface widening (read-only `observedCount` for forward-compat) is deferred —
no current consumer needs it, no point speculating.
