+++ packages/admin/src/editor/SlashMenuPanel.test.tsx
@@ -1,51 +1,18 @@
-import { act, cleanup, render, screen } from "@testing-library/react";
+import { cleanup, render, screen } from "@testing-library/react";
 import { userEvent } from "@testing-library/user-event";
-import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
+import { afterEach, describe, expect, test, vi } from "vitest";
 
 import { createBlockRegistry, createPatternRegistry } from "@plumix/blocks";
 
 import type { SlashMenuItem } from "./slash-menu-items.js";
+import { installFakeIntersectionObserver } from "./intersection-observer-harness.js";
 import { SlashMenuPanel } from "./SlashMenuPanel.js";
 
-interface FakeObserverHandle {
-  readonly callback: IntersectionObserverCallback;
-}
-
-let observers: FakeObserverHandle[];
-
-beforeEach(() => {
-  observers = [];
-  class FakeObserver {
-    constructor(public readonly callback: IntersectionObserverCallback) {
-      observers.push({ callback });
-    }
-    observe = vi.fn();
-    unobserve = vi.fn();
-    disconnect = vi.fn();
-    takeRecords = vi.fn((): IntersectionObserverEntry[] => []);
-  }
-  vi.stubGlobal("IntersectionObserver", FakeObserver);
-});
+const { intersect } = installFakeIntersectionObserver();
 
 afterEach(() => {
   cleanup();
-  vi.unstubAllGlobals();
 });
-
-function intersect(): void {
-  const entry = {
-    isIntersecting: true,
-    intersectionRatio: 1,
-    target: document.createElement("div"),
-    boundingClientRect: {} as DOMRectReadOnly,
-    intersectionRect: {} as DOMRectReadOnly,
-    rootBounds: null,
-    time: 0,
-  } satisfies IntersectionObserverEntry;
-  act(() => {
-    for (const o of observers) o.callback([entry], {} as IntersectionObserver);
-  });
-}
 
 const blocks = createBlockRegistry([]);
 const patterns = createPatternRegistry([]);
